+++ src/XrmDefinitelyTyped/AssemblyInfo.fs
@@ -4,18 +4,18 @@ open System.Reflection
 
 [<assembly: AssemblyTitleAttribute("XrmDefinitelyTyped")>]
 [<assembly: AssemblyProductAttribute("XrmDefinitelyTyped")>]
-[<assembly: AssemblyDescriptionAttribute("Tool to generate TypeScript declaration files for MS CRM Dynamics client-side coding.")>]
+[<assembly: AssemblyDescriptionAttribute("Tool to generate TypeScript declaration files for MS Dynamics 365/CRM client-side coding.")>]
 [<assembly: AssemblyCompanyAttribute("Delegate A/S")>]
-[<assembly: AssemblyCopyrightAttribute("Copyright (c) Delegate A/S 2015")>]
-[<assembly: AssemblyVersionAttribute("2.0.7")>]
-[<assembly: AssemblyFileVersionAttribute("2.0.7")>]
+[<assembly: AssemblyCopyrightAttribute("Copyright (c) Delegate A/S 2017")>]
+[<assembly: AssemblyVersionAttribute("2.1.0")>]
+[<assembly: AssemblyFileVersionAttribute("2.1.0")>]
 do ()
 
 module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyTitle = "XrmDefinitelyTyped"
     let [<Literal>] AssemblyProduct = "XrmDefinitelyTyped"
-    let [<Literal>] AssemblyDescription = "Tool to generate TypeScript declaration files for MS CRM Dynamics client-side coding."
+    let [<Literal>] AssemblyDescription = "Tool to generate TypeScript declaration files for MS Dynamics 365/CRM client-side coding."
     let [<Literal>] AssemblyCompany = "Delegate A/S"
-    let [<Literal>] AssemblyCopyright = "Copyright (c) Delegate A/S 2015"
-    let [<Literal>] AssemblyVersion = "2.0.7"
-    let [<Literal>] AssemblyFileVersion = "2.0.7"
+    let [<Literal>] AssemblyCopyright = "Copyright (c) Delegate A/S 2017"
+    let [<Literal>] AssemblyVersion = "2.1.0"
+    let [<Literal>] AssemblyFileVersion = "2.1.0"