+++ src/XrmDefinitelyTyped/Generation/GenerationMain.fs
@@ -1,6 +1,6 @@
 ﻿module DG.XrmDefinitelyTyped.GenerationMain
 
-open System
+open System.IO
 open Utility
 
 open CrmBaseHelper
@@ -26,35 +26,49 @@ let retrieveRawState xrmAuth rSettings =
 let generateFromRaw gSettings rawState =
   let out = gSettings.out ?| "."
   let formIntersects = gSettings.formIntersects ?| [||]
-  let crmVersion =
-    gSettings.crmVersion ?| (Int32.MaxValue, Int32.MaxValue, Int32.MaxValue, Int32.MaxValue)
+  let crmVersion = gSettings.crmVersion ?| rawState.crmVersion
 
   // Pre-generation tasks 
   clearOldOutputFiles out
   generateFolderStructure out gSettings
-  copyJsLibResourceFiles gSettings
-  copyTsLibResourceFiles gSettings
 
-  // Generate the files
+  // Interpret data and generate resource files
   let data =
     interpretCrmData out formIntersects rawState 
-    |>> generateDtsResourceFiles crmVersion gSettings 
-    |>> generateEnumFiles 
-    |>> generateEntityEnumFiles 
 
-  match gSettings.skipForms with
-  | false -> generateFormFiles data
-  | true  -> ()
+  let defs = 
+    seq {
+      yield! generateEnumDefs data
+      if not gSettings.skipForms then yield! generateFormDefs data
 
-  match gSettings.webNs with
-  | Some ns -> 
-    generateBaseWebEntityFile ns data
-    generateWebEntityFiles ns data
-  | None -> ()
+      match gSettings.webNs with
+      | Some ns -> 
+        yield generateBaseWebEntityDef ns data
+        yield! generateWebEntityDefs ns data
+      | None -> ()
 
-  match gSettings.restNs with
-  | Some ns -> 
-    generateRestFile ns data
-    generateBaseRestEntityFile ns data
-    generateRestEntityFiles ns data
-  | None -> ()
+      match gSettings.restNs with
+      | Some ns -> 
+        yield generateRestDef ns data
+        yield generateBaseRestEntityDef ns data
+        yield! generateRestEntityDefs ns data
+      | None -> ()
+    }
+    |> Array.ofSeq
+  
+  printf "Writing to files..."
+  copyJsLibResourceFiles gSettings
+  copyTsLibResourceFiles gSettings
+  generateDtsResourceFiles crmVersion gSettings data
+
+  match gSettings.oneFile with
+  | false -> 
+    defs 
+    |> Array.Parallel.iter (fun (path, lines) -> 
+      Directory.CreateDirectory path |> ignore
+      File.WriteAllLines(path, lines)
+    )
+  | true  -> 
+    let singleFilePath = Path.Combine(out, "context.d.ts")
+    defs |> Array.Parallel.map snd |> List.concat |> fun lines -> File.WriteAllLines(singleFilePath, lines)
+  printfn "Done!"